Delayed creation tweaks
Fixed an issue where creating a ColorPicker object too early would result in making a half-finished ColorPicker object, abandoning it, and creating a new one entirely on CreateQueuedMenus() instead of finishing the creation of the first object
Now, ColorPicker will return the object on creation, though it will not be fully set up. The boolean "init_done" in a ColorPicker object, which indicates that the object is fully created, should be verified as "true" before attempting to use the ColorPicker.